FreqDiffFormer/requirements.txt
torch>=1.12
torchvision
timm
numpy
scipy
PyYAML
opencv-python
scikit-learn
scikit-image
tqdm
matplotlib
einops
